AFK foundry auto-crafter
With this script you can walk away from your keyboard and it will keep crafting those pesky 1-minute 10x consumable recipes in your foundry (i.e. derelict keys, 10x energy restores, 10x air support charges, 10x ciphers, etc.)
note: This is made for 1920x1080 res. If your resolution is different, this will not work for you.

DISCLAIMER: Use this at your own risk. I've optimized the script to be as safe as possible in afk mode, so in ideal afk conditions this script should never use plat for rushing; accidental or otherwise. However, if this does happen, due to external stimulus or otherwise, I am not responsible for the plat you lose due to rushing. I will not compensate you, and I will not be held accountable for your decision - the act of using this script is yours and yours alone.

Now that the disclaimer out of the way, here's how you use it:
- Download and install AutoHotkey (google it).
- Open notepad, copy this script into it, and save it as "autocrafter.ahk" onto your d